requirements.txt
streamlit>=1.33.0
openai>=1.42.0
faiss-cpu>=1.8.0
pymupdf>=1.24.10
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
orjson>=3.10.0
